Removed duplicate objects from Scene2 and 3. Changed SceneManager to SceneManagerClass
