+++ Bricks/Assets/Scenes/Level1.unity
@@ -4809,10 +4809,10 @@ Transform:
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 519420028}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: -2.71}
+  m_LocalPosition: {x: -0.9085349, y: 4.4455028, z: -2.1}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
-  m_Father: {fileID: 0}
+  m_Father: {fileID: 1250769973}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &525014329
@@ -5053,6 +5053,50 @@ Transform:
   m_Father: {fileID: 1947761679}
   m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 90}
+--- !u!1 &1250769971
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1250769973}
+  - component: {fileID: 1250769972}
+  m_Layer: 0
+  m_Name: SceneController
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1250769972
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1250769971}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3a432ff4248353b4284ead3b09c5dfa9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  txtscore: {fileID: 2065967767}
+  globalScore: 0
+--- !u!4 &1250769973
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1250769971}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 0.90853477, y: -4.445503, z: -0.61}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children:
+  - {fileID: 519420032}
+  - {fileID: 1531895368}
+  m_Father: {fileID: 0}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1341675451
 GameObject:
   m_ObjectHideFlags: 0
@@ -5266,8 +5310,8 @@ RectTransform:
   m_Children:
   - {fileID: 1703059972}
   - {fileID: 2065967766}
-  m_Father: {fileID: 1947761679}
-  m_RootOrder: 4
+  m_Father: {fileID: 1250769973}
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -5611,7 +5655,6 @@ Transform:
   - {fileID: 1840002358}
   - {fileID: 1905730724}
   - {fileID: 697351012}
-  - {fileID: 1531895368}
   m_Father: {fileID: 0}
   m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -5658,7 +5701,6 @@ GameObject:
   m_Component:
   - component: {fileID: 2065967766}
   - component: {fileID: 2065967768}
-  - component: {fileID: 2065967769}
   - component: {fileID: 2065967767}
   m_Layer: 5
   m_Name: Text
@@ -5682,7 +5724,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -308, y: 230}
+  m_AnchoredPosition: {x: -294, y: 210}
   m_SizeDelta: {x: 160, y: 30}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &2065967767
@@ -5725,19 +5767,6 @@ CanvasRenderer:
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 2065967765}
   m_CullTransparentMesh: 0
---- !u!114 &2065967769
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 2065967765}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3a432ff4248353b4284ead3b09c5dfa9, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  txtscore: {fileID: 2065967767}
-  globalScore: 0
 --- !u!1 &2068201486
 GameObject:
   m_ObjectHideFlags: 0
