+++ Bricks/ProjectSettings/EditorBuildSettings.asset
@@ -9,7 +9,7 @@ EditorBuildSettings:
     path: Assets/Scenes/Level1.unity
     guid: 2cda990e2423bbf4892e6590ba056729
   - enabled: 1
-    path: Assets/Scenes/Level 2.unity
+    path: Assets/Scenes/Level2.unity
     guid: 2ddc4f44cc9613c489577dd33ea437d9
   - enabled: 1
     path: Assets/Scenes/Level3.unity
