+++ Bricks/ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 2018.2.5f1
+m_EditorVersion: 2018.2.2f1
